requirements.txt
pillow>=9.3.0
python-vlc>=3.0.0